==> code/00 Data Cleaning.ipynb <==
# Data Cleaning

## Formatting
Classical Chinese *juéjù* 絕句 or quatrains have either five or seven syllable lines for a total of 20 or 28 characters per line. The json text corpus includes other types of poems, with commas and periods in their formatting. 

## Selecting the Data
To select only quatrains, the number of lines are evaluated and quatrains are then stored in their respective pentasyllabic and heptasyllabic dataframes.


### Imports

import numpy as np
import pandas as pd
import os
import re

# Create a files variable that contains all of our data files.
files = os.listdir('../json')

df = pd.read_json('../json/poet.tang.0.json')

df.head()

# Function that checks if a poem_raw is a quatrain
def quatrain_checker(poem):
    if len(poem)!=2:
        return False
    poem_raw = poem.copy()
    poem_raw[0] = re.sub('\ |\?|\？|—|\《|\》|\□|\●|\/|\{|\}|\·|\、|\「|\」|\|', '' , poem_raw[0])
    poem_raw[1] = re.sub('\ |\?|\？|—|\《|\》|\□|\●|\/|\{|\}|\·|\、|\「|\」|\|', '' , poem_raw[1])
    poem_raw[0] = poem_raw[0].split('，')
    poem_raw[1] = poem_raw[1].split('，')
    
    if len(poem_raw[0])!=2:
        return False
    if len(poem_raw[1])!=2:
        return False
    try:
        line_lengths = [len(poem_raw[0][0]), len(poem_raw[0][1]), len(poem_raw[1][0]), len(poem_raw[1][1])]
    except:
        return False
    
    if line_lengths == [5, 6, 5, 6]:
        return True
    elif line_lengths == [7, 8, 7, 8]:
        return True
    else:
        return False

# Function that takes a .json file and returns DataFrame with only quatrains labeled by amount of syllables per line
def quatrain_extractor(file):
    poetry = pd.read_json(f'../json/{file}')
    cleaned_poetry = []
    for i in range(0, len(poetry.paragraphs)):
        poem = poetry.paragraphs[i]
        if quatrain_checker(poem) == True:            
            poem = re.sub('\，|\。', '', ''.join(poem))
            if len(poem)==20:
                cleaned_poetry.append({'text': poem, 'syllables': 5})
            elif len(poem)==28:
                cleaned_poetry.append({'text': poem, 'syllables': 7})
    return pd.DataFrame(cleaned_poetry)

# Creating DataFrame with all quatrains
poetry = pd.concat([quatrain_extractor(n) for n in files], ignore_index=True)

poetry.syllables.value_counts()

len(poetry)

len(poetry.text.unique())

# Removing duplicates
poetry.drop_duplicates(subset='text', inplace=True)

# Confirming duplicate deletion
len(poetry)

poetry.syllables.value_counts()

# Resetting index
poetry.reset_index(inplace=True, drop=True)

# Exporting to CSV
poetry.to_csv('../data/quatrains.csv')
