==> code/01 Vectorizer.ipynb <==
# Vectorizer

## Generating Words List
Using a simple count vectorizer, a list of unique characters were extracted for further use. Previous iterations of this notebook were also used to find punctuation marks missing from poem cleaning regex expression and further clean the list of quatrain poems.

### Imports

import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer

df = pd.read_csv('../data/quatrains.csv')

df['text_split'] = df.text.map(lambda x: ' '.join(x))

X = df['text_split']

cvec = CountVectorizer(tokenizer=lambda x: x.split())

X = cvec.fit_transform(X)

len(pd.DataFrame(cvec.get_feature_names())[0])

char_list = pd.DataFrame(cvec.get_feature_names())

char_list

char_list.to_excel('../data/char_list.xlsx')
