==> code/02 Scraper and Dictionary Building.ipynb <==
# Character Dictionary Creation

## Formatting
Several pieces of metadata for each character are crucial to the identification of rhyme and meter, and that is each characters rime and tone class. For the purposed of Classical Chinese poetry, tones classes are divided into 平 level tones and 仄 oblique tones, the latter contains every character that isn't level tone. These are respectively indicated as 1 and 0 in the data.

## Further Data Annotation
Pronunciation in modern languages given for ease of reference

### Imports

import numpy as np
import pandas as pd
from bs4 import BeautifulSoup
import requests
import pinyin #For Mandarin
from pyjyutping import jyutping #For Cantonese
import hanja #For Korean

def char_info(char):
    info = {'char': char}
    response = requests.get(f'https://ytenx.org/zim?dzih={char}&dzyen=1&jtkb=1&jtkd=1&jtdt=1&jtgt=1')
    try: 
        soup = BeautifulSoup(response.text, 'lxml')
        entry = list(soup.find_all('p')[0])
        if entry[0] == '\n':
            return pd.DataFrame(info, index=[0])
        info['tone'] = entry[0][-1]
        info['rime'] = entry[1].text
        if info['tone'] == '入': #入 or entering tone characters were phonemically toneless but were considering a tone class for poetry purposes
            info['ipa'] = entry[4].text.split()[1][0:-1] #these characters end with stop consonants -p, -t, -k
            info['tone_class'] = 'E'
        elif info['tone'] == '平': #平 tones were level tones
            info['ipa'] = entry[4].text.split()[1][0:-1]
            info['tone_class'] = 'L'        
        else:
            info['ipa'] = entry[4].text.split()[1][0:-2] #去 departing and 上 rising tones are marked with H and X respectively in many reconstructions and are attached to the IPA in this database
            info['tone_class'] = entry[4].text.split()[1][-2]
        return pd.DataFrame(info, index=[0])
    except:
        return pd.DataFrame()

char_info('明')

char_info('起')

char_info('會')

char_info('烈')

df = pd.read_csv('../data/char_list.csv', index_col=0)

df

%%time
dictionary = pd.concat([char_info(i) for i in df.char])
#getting metadata for every character

#counting nulls and checking types
dictionary.info()

dictionary.to_csv('/../data/dictionary.csv')

#adding Mandarin pronunciation
dictionary['pinyin'] = dictionary['char'].map(lambda x: pinyin.get(x))

#adding Cantonese pronunciation
dictionary['jyutping'] = dictionary['char'].map(lambda x: jyutping.convert(x))

#adding Korean pronunciation
dictionary['hangul'] = dictionary['char'].map(lambda x: hanja.translate(x, 'substitution'))

dictionary.iloc[1000:1030,:]

dictionary.to_csv('../data/dictionary_expanded.csv')

#dropping nulls
clean_dictionary = pd.read_csv('../data/dictionary_expanded.csv.csv', index_col=0)
clean_dictionary = clean_dictionary.dropna()
clean_dictionary.reset_index(drop=True, inplace=True)

clean_dictionary['rime'] = clean_dictionary['rime'].map(lambda x: x[0])

#creating initial rime index
rimes = list(clean_dictionary.rime.unique())
rimes = pd.DataFrame(rimes, index=range(0,208), columns=['rime'])
rimes.to_csv('../../data/rimes.csv', index=None)

rimes['rime_index'] = rimes.index

rimes[rimes['rime'] == '送']

#adding index to all chars
clean_dictionary = clean_dictionary.join(rimes.set_index('rime'), on='rime', how='left')

#adding meter class {1 for L 平 level tones, 0 for all other tones as oblique tones}
clean_dictionary['meter_class'] = clean_dictionary['tone_class'].map(lambda x: 1 if x=='L' else 0)

clean_dictionary.iloc[1000:1010,:]

rime_table = pd.read_csv('../data/guangyun_rime_table.csv')

rime_index = [{rime_table.E[i]: rime_table.rime_index[i]} for i in range(0,79) if type(rime_table.E[i])==str] + [{rime_table.L[i]: rime_table.rime_index[i]} for i in range(0,79) if type(rime_table.L[i])==str] + [{rime_table.X[i]: rime_table.rime_index[i]} for i in range(0,79) if type(rime_table.X[i])==str] + [{rime_table.H[i]: rime_table.rime_index[i]} for i in range(0,79) if type(rime_table.H[i])==str]

rime_indices = {}
for i in rime_index:
    rime_indices.update(i)

rime_table.to_csv('../../data/guangyun_rime_table.csv')

#resetting rime index to catch errors missing rimes
clean_dictionary['rime_index'] = 999

#remapping rime index
clean_dictionary['rime_index'] = clean_dictionary.rime.map(rime_indices)

#checking for errors
clean_dictionary[clean_dictionary['rime_index']==999]

clean_dictionary.iloc[1000:1010,:]

#exporting final dictionary
clean_dictionary.to_csv('../data/dictionary_cleaned.csv')
