==> code/03 Poetry Cleaning.ipynb <==
# Second Round Poetry Cleaning

## Formatting
Because a computer cannot evaluate poetry that that contains characters it doesn't know, those poems will need to be removed. Most of these are variant characters, but since I couldn not find a quality variant database, there was no easy way for the computer to map the variant character to a corresponding standard character

## Selecting the Data
Heptasyllabic poems greatly outnumber pentasyllabic ones so models will focus on them.

### Imports

import numpy as np
import pandas as pd

poetry = pd.read_csv('../data/quatrains.csv', index_col=0)

poetry

# removing non heptasyllabic poems
poetry = poetry[poetry.syllables == 7]

poetry

#reading in dictionary
dictionary = pd.read_csv('../data/dictionary_cleaned.csv', index_col=0)

dictionary

# creating character list for ease of access
char_list = [i for i in dictionary.char]
char_list = set(char_list)

# creating character metadata dictionary for ease of access
chars = {dictionary.char[i]: i for i in range(0,len(dictionary))}
char_dict = {i: {'char': dictionary.char[i],
                 'tone': dictionary.tone_class[i],
                 'rime': dictionary.rime_index[i],
                 'meter': dictionary.meter_class[i]
                 } for i in range(0,len(dictionary))}

# function that checks if a poem contains invalid characters
def character_checker(poem):
    invalid_count = 0
    for i in poem:
        if i not in char_list:
            invalid_count += 1
    if invalid_count == 0:
        return True
    else:
        return False            

# function that takes in a poem string and outputs a list of character tokens
def text_tokenizer(text):
    return [chars[i] for i in text]

# function that takes in a list of tokens and determines if they rhyme or not
def rhyme_checker(tokens):
    if char_dict[tokens[13]]['rime'] == char_dict[tokens[27]]['rime']:
        return True
    else:
        return False

# functions that takes in poem string returns true if there are no invalid characters and the poem rhymes
def poem_check(poem):
    if character_checker(poem) == True and rhyme_checker(text_tokenizer(poem)) == True:
        return True
    else:
        return False

poetry['valid'] = poetry['text'].apply(poem_check)

# checking the amount of valid poems 
poetry.valid.value_counts()

# removing invalid poems 
valid_poetry = poetry[poetry['valid']==True]

valid_poetry.reset_index(drop=True, inplace=True)

# exporting to CSV
valid_poetry['text'].to_csv('../data/final_quatrains.csv', index=0)

valid_poetry
