==> code/04 GAN Modeling.ipynb <==
# GAN Model

### Imports

import pandas as pd
import numpy as np

poetry = pd.read_csv('../data/final_quatrains.csv')

dictionary = pd.read_csv('../data/dictionary_cleaned.csv', index_col=0)

poetry

dictionary

chars = {dictionary.char[i]: i for i in range(0,len(dictionary))}
char_dict = {i: {'char': dictionary.char[i],
                 'tone': dictionary.tone_class[i],
                 'rime': dictionary.rime_index[i],
                 'meter': dictionary.meter_class[i]
                 } for i in range(0,len(dictionary))}

meters = [
    [1,1,0,0,1,1,0,
     0,0,1,1,0,0,1,
     0,0,1,1,0,0,1,
     1,1,0,0,0,1,1],
    [0,0,1,1,1,0,0,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,0,1,1,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1],
    [0,0,1,1,0,0,1,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [0,0,1,1,0,0,1,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,0,1,1,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1],
    [0,0,1,1,1,0,0,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,1,1,0,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1]
]

# function that takes in a poem string and outputs a list of character tokens
def text_tokenizer(text):
    return [chars[i] for i in text]

# function that takes in a list of character tokens and outputs a poem string
def token_to_text(tokens):
    return ''.join([char_dict[i]['char'] for i in tokens])

# function that takes in a list of tokens and determines if they rhyme or not
def rhyme_checker(tokens):
    if char_dict[tokens[13]]['rime'] == char_dict[tokens[27]]['rime']:
        return 1
    else:
        return 0

# show meter classes of all tokens as a list
def show_meter(tokens):
    return [char_dict[i]['meter'] for i in tokens]

# calculates meter adherence and outputs highest meter score
def meter_score(tokens):
    meter = [char_dict[i]['meter'] for i in tokens]
    scores = []
    for n in meters:
        scores.append(sum([int(meter[num]==n[num]) for num in n]))
    return max(scores)/28

# creating column for tokens
poetry['tokens'] = poetry['text'].apply(text_tokenizer)

# creating column for meter score
poetry['meter_score'] = poetry['tokens'].apply(meter_score)

# creating column for rhyme
poetry['rhyme'] = poetry['tokens'].apply(rhyme_checker)

# creating positive class for all real poems for modeling
poetry['target'] = 1

# saving dataframe for further use
poetry.to_csv('../data/quatrains_w_info.csv')

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, LSTM, Embedding, Bidirectional
from sklearn.model_selection import train_test_split

tf.random.set_seed(42)
np.random.seed(42)

# adds meter and rhyme checking to text for discriminator
def text_append(text):
    tokenized = text_tokenizer(text)
    j = meter_score(tokenized)
    k = rhyme_checker(tokenized)
    tokenized.append(j)
    tokenized.append(k)
    return tokenized

X

# building discriminator model
discriminator = Sequential()
discriminator.add(Embedding(input_dim=voc_size, output_dim=64))
discriminator.add(Bidirectional(LSTM(64)))
discriminator.add(Dense(128, activation='relu'))
discriminator.add(Dense(64, activation='relu'))
discriminator.add(Dense(1))

discriminator.compile(loss='binary_crossentropy', optimizer='adam')

# building generator model
coding_size = 100
seq_len = 28
voc_size = 7980
generator = Sequential()
generator.add(Embedding(voc_size+1, seq_len))
generator.add(LSTM(140, return_sequences=True))
generator.add(LSTM(70))
generator.add(Dense(56, activation='relu'))
generator.add(Dense(seq_len, activation='sigmoid'))
# no need to compile

GAN = Sequential([generator, discriminator])

# turn off discriminator training
discriminator.trainable = False

# compile GAN
GAN.compile(loss='binary_crossentropy', optimizer='adam')

batch_size = 267

# convert Dataframe to TensorFlow Dataset
dataset = tf.data.Dataset.from_tensor_slices(X.to_dict())

dataset = dataset.batch(batch_size, drop_remainder=True).prefetch(1)

epoch = 200

# testing generator
noise = tf.random.normal(shape=[batch_size, voc_size+1])
gen_poems = generator(noise)

gen_poems

# GAN modified from Jose Portilla's 'Tensorflow 2 and Keras Deep Learning' course
# Grab the seprate components
generator, discriminator = GAN.layers

# For every epcoh
for epoch in range(epochs):
    print(f"Currently on Epoch {epoch+1}")
    i = 0
    # For every batch in the dataset
    for X_batch in dataset:
        i=i+1
        if i%100 == 0:
            print(f"\tCurrently on batch number {i} of {len(my_data)//batch_size}")

        ## TRAINING THE DISCRIMINATOR ######
        # Create Noise
        noise = tf.random.normal(shape=[batch_size, codings_size])
        
        # Generate numbers based just on noise input
        gen_poems = generator(noise)
        
        # Append metrics to fake poems so discriminator has more data to work off of
        meter_score(gen_poems)
        rhyme_checker(gen_poems)     
        
        # Concatenate Generated Text against the Real Ones
        X_fake_vs_real = tf.concat([gen_poems, tf.dtypes.cast(X_batch,tf.float32)], axis=0)
        
        # Targets set to zero for fake text and 1 for real text
        y1 = tf.constant([[0.]] * batch_size + [[1.]] * batch_size)
        
        # This gets rid of a Keras warning
        discriminator.trainable = True
        
        # Train the discriminator on this batch
        discriminator.train_on_batch(X_fake_vs_real, y1)
        
        ## TRAINING THE GENERATOR     ######

        # Create some noise
        noise = tf.random.normal(shape=[batch_size, codings_size])
        
        # We want discriminator to belive that fake text are real
        y2 = tf.constant([[1.]] * batch_size)
        
        # Avoids a warning
        discriminator.trainable = False
        
        GAN.train_on_batch(noise, y2)
        
print("TRAINING COMPLETE")
