==> code/05 Generation Concept with LSTMs.ipynb <==
# LSTM-based Text Prediction

## Methodology
Using a more basic neural network, this model takes in 7 characters and predicts and 8th

### Imports

import pandas as pd
import numpy as np

poetry = pd.read_csv('../data/quatrains_w_info.csv', index_col=0)

dictionary = pd.read_csv('../data/dictionary_cleaned.csv', index_col=0)

poetry

dictionary.iloc[4000:4015,:]

chars = {dictionary.char[i]: i for i in range(0,len(dictionary))}
char_dict = {i: {'char': dictionary.char[i],
                 'tone': dictionary.tone_class[i],
                 'rime': dictionary.rime_index[i],
                 'meter': dictionary.meter_class[i]
                 } for i in range(0,len(dictionary))}

meters = [
    [1,1,0,0,1,1,0,
     0,0,1,1,0,0,1,
     0,0,1,1,0,0,1,
     1,1,0,0,0,1,1],
    [0,0,1,1,1,0,0,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,0,1,1,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1],
    [0,0,1,1,0,0,1,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [0,0,1,1,0,0,1,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,0,1,1,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1],
    [0,0,1,1,1,0,0,
     1,1,0,0,0,1,1,
     1,1,0,0,1,1,0,
     0,0,1,1,0,0,1],
    [1,1,0,0,1,1,0,
     0,0,1,1,0,0,1,
     0,0,1,1,1,0,0,
     1,1,0,0,0,1,1]
]

# function that takes in a poem string and outputs a list of character tokens
def text_tokenizer(text):
    return [chars[i] for i in text]

# function that takes in a list of character tokens and outputs a poem string
def token_to_text(tokens):
    return ''.join([char_dict[i]['char'] for i in tokens])

# function that takes in a list of tokens and determines if they rhyme or not
def rhyme_checker(tokens):
    if char_dict[tokens[13]]['rime'] == char_dict[tokens[27]]['rime']:
        return 1
    else:
        return 0

# show meter classes of all tokens as a list
def show_meter(tokens):
    return [char_dict[i]['meter'] for i in tokens]

# calculates meter adherence and outputs highest meter score
def meter_score(tokens):
    meter = [char_dict[i]['meter'] for i in tokens]
    scores = []
    for n in meters:
        scores.append(sum([int(meter[num]==n[num]) for num in n]))
    return max(scores)/28

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense,LSTM,Embedding
from tensorflow.keras.utils import to_categorical

tf.random.set_seed(42)
np.random.seed(42)

# function that 
def sequencer(tokens, train_len):
    sequences = []
    for i in range(train_len, len(tokens)):
        seq = tokens[i-train_len: i]
        sequences.append(seq)
    return sequences

sequence_list = []
for i in poetry.tokens:
    sequence_list += sequencer(i, 8)

len(sequence_list)

sequence_list = np.array(sequence_list)

vocabulary_size = len(chars)

len(chars)

X = sequence_list[:,:-1]
y = sequence_list[:,-1]

y = to_categorical(y, num_classes=vocabulary_size)

seq_len = X.shape[1]

X.shape

X[0]

model = Sequential()
model.add(Embedding(vocabulary_size, seq_len, input_length=seq_len))
model.add(LSTM(28, return_sequences=True))
model.add(LSTM(70))
model.add(Dense(49, activation='relu'))

model.add(Dense(vocabulary_size, activation='softmax'))
    
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
   
model.summary()

model.fit(X, y, batch_size=1028, epochs=200, verbose=1)

from tensorflow.keras.models import load_model
model = load_model('../models')

texts = ['四望雲天直下低', '牽絲作老翁雞皮']
texts = [text_tokenizer(x) for x in texts]
texts = np.array(texts)

preds = model.predict(texts)

type(np.argmax(preds[0]))

np.argmax(preds[1])

type(int(preds))

char_dict[74]['char']

char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char'], char_dict[np.random.randint(0, vocabularly_size)]['char']

def poem_generator(model, seed_text):
    output_text = [] + text_tokenizer(seed_text)
    for i in range()
